create controllers, add status, history

+++ letter-service/src/main/java/com/test/postservice/letterservice/service/PostItemServiceImpl.java
@@ -1,16 +1,18 @@
 package com.test.postservice.letterservice.service;
 
-import com.test.postservice.letterservice.dto.PostItemDto;
 import com.test.postservice.letterservice.entity.PostItem;
-import com.test.postservice.letterservice.mapper.PostItemMapper;
+import com.test.postservice.letterservice.entity.PostItemStatus;
+import com.test.postservice.letterservice.entity.PostOffice;
 import com.test.postservice.letterservice.repository.PostItemRepository;
 import jakarta.ws.rs.BadRequestException;
+import jakarta.ws.rs.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -20,22 +22,19 @@ public class PostItemServiceImpl implements PostItemService {
 
     private final PostItemRepository postItemRepository;
 
-    public void update(PostItem postItem) {
-        postItemRepository.save(postItem);
+    @Override
+    public List<PostItem> findAll() {
+        return postItemRepository.findAll();
     }
 
     @Override
-    public List<PostItemDto> findAll() {
-        return List.of();
+    public PostItem findById(long id) {
+        log.info("item with ID{}", id);
+        return postItemRepository.findById(id).orElseThrow(() -> new NotFoundException(String.format("Item with id %d not found", id)));
     }
 
-    @Override
-    public PostItemDto findById(long id) {
-        return null;
-    }
-
-    @Override
     @Transactional
+    @Override
     public boolean create(PostItem postItem) {
         Optional<PostItem> postItemOptional = postItemRepository.findById(postItem.getId());
         if (postItemOptional.isPresent()) {
@@ -43,17 +42,50 @@ public class PostItemServiceImpl implements PostItemService {
             throw new BadRequestException(String.format("post item with id %d already exists",
                     postItem.getId()));
         }
+        postItem.setId(null);
+        postItem.setStatus(PostItemStatus.ACCEPTED);
         postItemRepository.saveAndFlush(postItem);
-        log.info("create atm {}", postItem);
+        log.info("create item {}", postItem);
         return true;
+    }
+    @Transactional
+    @Override
+    public boolean update(PostItem postItem) {
+        Optional<PostItem> postItemOptional = postItemRepository.findById(postItem.getId());
+        if (postItemOptional.isPresent()) {
+            postItemRepository.save(postItem);
+            log.info("update item with ID{}", postItem.getId());
+            return true;
+        }
+        log.error("update item error, ID{}", postItem.getId());
+        return false;
+    }
+    @Transactional
+    @Override
+    public boolean delete(Long id) {
+        if (postItemRepository.findById(id).isPresent()) {
+            postItemRepository.deleteById(id);
+            log.info("delete item with ID{}", id);
+            return true;
+        }
+        log.error("delete item error, with id{}", id);
+        return false;
     }
 
     @Override
-    public PostItemDto update(Long id, PostItemDto accountTransferDto) {
-        return null;
-    }
+    public boolean setPostOffices(PostOffice postOffice, Long id) {
+        Optional<PostItem> postItem = postItemRepository.findById(id);
+        if (postItem.isEmpty()) {
+            return false;
+        }
 
-    public void delete(Long id) {
-        postItemRepository.deleteById(id);
+        List<PostOffice> postOfficeList = postItem.get().getPostOffices();
+        postOfficeList.add(postOffice);
+        postItem.get().setPostOffices(postOfficeList);
+        if (!Objects.equals(postItem.get().getPostCode(), postOffice.getPostCode())) {
+            postItem.get().setStatus(PostItemStatus.SEND_TO_INTERMEDIATE);
+        }
+        postItemRepository.saveAndFlush(postItem.get());
+        return true;
     }
 }
